+++ models.py
@@ -1,0 +1,1 @@
+from django.db import models
