+++ myapp/forms.py
@@ -1,0 +1,6 @@
+from django import forms
+
+class ContactForm (forms.Form):
+        nombre = forms.CharField(max_length=100)
+        correo = forms.EmailField()
+        mensaje = forms.CharField(widget=forms.Textarea)
