+++ myapp/templates/myapp/legal.html
