+++ myapp/templates/myapp/static/js/theme.js
