--- mysite/forms.py
@@ -1,7 +1,0 @@
-from django import forms
-from .models import ContactMessage
-
-class ContactForm(forms.Form):
-    nombre = forms.CharField(max_length=100)
-    correo = forms.EmailField()
-    mensaje = forms.CharField(widget=forms.Textarea)
