+++ mysite/urls.py
@@ -15,15 +15,10 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import path
-from . import views
-from .views import contact_view
+from django.urls import path , include
 
 
 urlpatterns = [
-    path('', views.hello_world, name='hello_world'),
-    path('about/', views.about, name='about'),
-    path('legal/', views.legal, name='legal'),
-    path('contacto/', contact_view, name='contacto'),
-    path('admin/', admin.site.urls, name='admin'),
+    path('admin', admin.site.urls, name='admin'),
+    path('', include('myapp.urls'))
     ]
