+++ requiriments.txt
@@ -1,0 +1,9 @@
+asgiref==3.8.1
+distlib==0.3.8
+Django==5.1.2
+filelock==3.13.1
+platformdirs==4.2.0
+pytube==15.0.0
+sqlparse==0.5.1
+tzdata==2024.2
+virtualenv==20.25.1
